feat(11-react-music): Add @ant-design/v5-patch-for-react-19 and modify message and player - bar functionality
The commit adds the @ant - design/v5 - patch - for - react - 19 package to the project. Also, it modifies the common.less file to style the ant - message component. In the player - bar component, it imports and uses the 'message' from antd to show lyrics when clicked. This was likely done to enhance the user experience and potentially fix compatibility issues with React 19.